1、修改bug

+++ app/src/main/java/com/huojumu/model/WorkInfo.java
@@ -104,67 +104,67 @@ public class WorkInfo {
          * real : 15
          */
 
-        private int mau;
-        private int wx;
-        private int total;
-        private int virtual;
-        private int alp;
-        private int coup;
-        private int real;
+        private double mau;
+        private double wx;
+        private double total;
+        private double virtual;
+        private double alp;
+        private double coup;
+        private double real;
 
-        public int getMau() {
+        public double getMau() {
             return mau;
         }
 
-        public void setMau(int mau) {
+        public void setMau(double mau) {
             this.mau = mau;
         }
 
-        public int getWx() {
+        public double getWx() {
             return wx;
         }
 
-        public void setWx(int wx) {
+        public void setWx(double wx) {
             this.wx = wx;
         }
 
-        public int getTotal() {
+        public double getTotal() {
             return total;
         }
 
-        public void setTotal(int total) {
+        public void setTotal(double total) {
             this.total = total;
         }
 
-        public int getVirtual() {
+        public double getVirtual() {
             return virtual;
         }
 
-        public void setVirtual(int virtual) {
+        public void setVirtual(double virtual) {
             this.virtual = virtual;
         }
 
-        public int getAlp() {
+        public double getAlp() {
             return alp;
         }
 
-        public void setAlp(int alp) {
+        public void setAlp(double alp) {
             this.alp = alp;
         }
 
-        public int getCoup() {
+        public double getCoup() {
             return coup;
         }
 
-        public void setCoup(int coup) {
+        public void setCoup(double coup) {
             this.coup = coup;
         }
 
-        public int getReal() {
+        public double getReal() {
             return real;
         }
 
-        public void setReal(int real) {
+        public void setReal(double real) {
             this.real = real;
         }
     }
